backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
prophet
xgboost
python-multipart
plotly
requests
textblob
google-genai
python-dotenv
opencv-python
#torch
#torchvision
ultralytics
tensorflow
supervision
geopandas
shapely
osmnx
networkx
torch_geometric
shap
matplotlib
seaborn
tqdm
sentence-transformers
scikit-learn